VMC: Uart patch Clock Shutdown
- Set the LCR/Parity only for Uart0 (SC Comms).
- Uart Receive timeout is now configurable.
- Added Clock shutdown feature, and threshold monitoring for VCCINT_Temp.
- Fetch VCCINT_Temp from MSP over Uart.
- Added Active and Target MSP version sensors in ASDM BoardInfo.
- Minor fixes.

+++ src/vmc/sensors/inc/se98a.h
@@ -8,6 +8,6 @@
 
 #include "xil_types.h"
 
-u8 SE98A_ReadTemperature(u8 i2c_num, u8 slaveAddr, s32 *temperatureValue);
+u8 SE98A_ReadTemperature(u8 i2c_num, u8 slaveAddr, s16 *temperatureValue);
 
 #endif
